added logic for registration and authentification, main routes are now private

+++ controllers/contactController.js
@@ -3,17 +3,17 @@ const Contact = require('../models/contactModel');
 
 const uri = process.env.URI;
 
-
+// all the routes below are private:
 const getAllContacts = async function(req, res) {
     try {
         await mongoose.connect(uri);
         console.log('connected');
 
         const contacts = await Contact.find();
-        res.json(contacts);
-        console.log(contacts);
+        res.status(200).json({ message: "the following contacts exist in a database atm:", contacts });
     } catch(err) {
         console.log(err);
+        return res.status(400).json({ error: err });
     } finally {
         await mongoose.disconnect();
         console.log('disconnected');
@@ -23,15 +23,19 @@ const getAllContacts = async function(req, res) {
 const postContact = async function(req, res) {
     const { name, email, phone } = req.body;
 
+    if (!name || !(email || phone)) {
+        return res.status(400).json( { message: "Please specify a name and either email or phone, or both!!"} );
+    }
+
     try {
         await mongoose.connect(uri);
         console.log('connected');
 
         const contact = await Contact.create( { name, email, phone });
-        res.json(contact);
-        console.log(contact);
+        res.status(200).json({ message: "contact posted", contact });
     } catch(err) {
         console.log(err);
+        return res.status(400).json({ error: err });
     } finally {
         await mongoose.disconnect();
         console.log('disconnected');
@@ -41,15 +45,19 @@ const postContact = async function(req, res) {
 const getSingleContact = async function(req, res) {
     const _id = req.params.id;
 
+    if (!_id) {
+        return res.status(400).json( { message: "Please fill in the id required!!"} );
+    }
+
     try {
         await mongoose.connect(uri);
         console.log('connected');
 
         const contact = await Contact.findById(_id);
-        res.json(contact);
-        console.log(contact);
+        res.status(200).json({ message: "contact found", contact });
     } catch(err) {
         console.log(err);
+        return res.status(400).json({ error: err });
     } finally {
         await mongoose.disconnect();
         console.log('disconnected');
@@ -59,15 +67,19 @@ const getSingleContact = async function(req, res) {
 const deleteSingleContact = async function(req, res) {
     const _id = req.params.id;
 
+    if (!_id) {
+        return res.status(400).json( { message: "Please fill in the id required!!"} );
+    }
+
     try {
         await mongoose.connect(uri);
         console.log('connected');
 
         const contact = await Contact.findByIdAndDelete( _id );
-        res.json({ message: "contact deleted", contact });
-        console.log(contact);
+        res.status(200).json({ message: "contact deleted", contact });
     } catch(err) {
         console.log(err);
+        return res.status(400).json({ error: err });
     } finally {
         await mongoose.disconnect();
         console.log('disconnected');
@@ -78,15 +90,23 @@ const updateSingleContact = async function(req, res) {
     const { name, email, phone } = req.body;
     const _id = req.params.id;
 
+    if (!(name || email || phone)) {
+        return res.status(400).json( { message: "Please specify the field you are going to update!!"} );
+    }
+
+    if (!_id) {
+        return res.status(400).json( { message: "Please fill in the id required!!"} );
+    }
+
     try {
         await mongoose.connect(uri);
         console.log('connected');
 
-        const contact = await Contact.findOneAndReplace( {_id}, { name, email, phone } );
-        res.json({ message: "contact replaced", contact });
-        console.log(contact);
+        const contact = await Contact.findOneAndUpdate( {_id}, { name, email, phone } );
+        res.status(200).json({ message: "contact replaced", contact });
     } catch(err) {
         console.log(err);
+        return res.status(400).json({ error: err });
     } finally {
         await mongoose.disconnect();
         console.log('disconnected');
